+++ blackmamba/bundle.py
@@ -8,7 +8,7 @@ from blackmamba.log import debug
 _BUNDLES = {
     'analyze': ['mccabe', 'flake8', 'pycodestyle', 'pyflakes', 'setuptools'],
     'docutils': ['docutils'],
-    'refactoring': ['rope']
+    'refactoring': ['rope', 'future', 'libfuturize', 'libpasteurize']
 }
 
 _BUNDLED_MODULES_PATH = os.path.abspath(os.path.join(os.path.dirname(os.path.abspath(__file__)), 'lib'))